[BUGFIX]iwlwifi: mvm: fix security bug in PN checking [SQUASH]
My previous patch allowed the same PN for packets originating
from the same AMSDU by copying PN only for the last packet in
the series.
This however is bogus since we cannot assume the last frame will
be received on the same queue, and if it is received on a different
ueue we will end up not incrementing the PN and possibly let
the next packet to have the same PN and pass through.
Change the logic instead to driver explicitly indicate for the
second sub frame and on to be allowed to have the same PN as
the first subframe. Indicate it to mac80211 as well for the
fallback queue.

type=bugfix
fixes=I13493187cb6b62aed1d2770798bd6d11d33ab5cc
bug=not-tracked

Change-Id: Ice87193dbf5faf16c99d8572619dd71ddd4452f2
Signed-off-by: Sara Sharon <sara.sharon@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/26069
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Grumbach, Emmanuel <emmanuel.grumbach@intel.com>
Reviewed-by: Berg, Johannes <johannes.berg@intel.com>
x-iwlwifi-stack-dev: cff870649dd92d0b9a40a539f4921db17f2e51f7