mac80211: fix RX u64 stats consistency on 32-bit platforms
On 32-bit platforms, the 64-bit counters we keep need to be protected
to be consistently read. Use the u64_stats_sync mechanism to do that.

In order to not end up with overly long lines, refactor the tidstats
assignments a bit.

type=cleanup
[again, we don't seem to care much, so let's not worry about bugfix]

Change-Id: I2090eb588cfdc36d0ca38fc52bd617fe5d8c8ec4
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/26126
Reviewed-by: Grumbach, Emmanuel <emmanuel.grumbach@intel.com>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
x-iwlwifi-stack-dev: 150dd7edbeb28538c662b4b18febc66612ebac03