iwlwifi: remove IWLWIFI_DEBUG_EXPERIMENTAL_UCODE
This Kconfig option allows to load a firmware for
debugging with a different name. This mechanism has not
been used for a few years now and replacing the firmware
file works as well.
Kill this Kconfig option and all the code that goes with it.

type=cleanup

Change-Id: I7c9ce631de846fc52fd09902c7e40a2617e9e119
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/27864
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Coelho, Luciano <luciano.coelho@intel.com>
x-iwlwifi-stack-dev: d575e1435b1f6446d0f7634c3ca132d4bc524617