[BUGFIX] iwlwifi: mvm: fix txq aggregation bug
Fix an issue where nullfunc frames and block ack requests
had the same tid as aggregation frames and were queued on
a non aggregation queue. The pending frames counter included
those frames but the check whether to decrement the pending
frames counter relied on the tid status and not on the txq id.
The result was an inconsistent state of the pending frames
counter followed by a failure to remove the station.
This failure triggered SYSASSERT 0x3421.

In addition, fix a situation in DQA mode where the number
of pending frames turned negative. This was due to the TX queue
being on the IWL_EMPTYING_HW_QUEUE_DELBA state and its frames
were still decremented.

Even though the SYSASSERT issue is fixed when DQA is disabled,
the issue is not completely solved when DQA is enabled and
should still be fixed.

type=bugfix
fixes=I35b4168139afd9c99d0607ba399ee99320f92f97
bug=cq:MWG100259462

Change-Id: I680e72b22a28e19c2845448126f9733813fe0d7e
Signed-off-by: Oren Givon <oren.givon@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/35657
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Coelho, Luciano <luciano.coelho@intel.com>
x-iwlwifi-stack-dev: 58c38127858bd724e37c700c7585b642ad6ea4ca