[BUGFIX] iwlwifi: pcie: enable interrupts before releasing the NIC's CPU
The NIC's CPU gets started after the firmware has been
written to its memory. The first thing it does is to
send an interrupt to let the driver know that it is
running. In order to get that interrupt, the driver needs
to make sure it is not masked. Of course, the interrupt
needs to be enabled in the driver before the CPU starts to
run.
I mistakenly inversed those two steps leading to races
which prevented the driver from getting the alive interrupt
from the firmware.
Fix that.

type=bugfix
bug=chrome-os-partner:53958
fixes=I68148ae7249772535a55e7722fd16200dd7bb2f8

Change-Id: I61fc49dd10998c06b5dc9ecba50ad4f3c869fff8
Fixes: a6bd005fe92 ("iwlwifi: pcie: fix RF-Kill vs. firmware load race") [4.5+]
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/37980
Reviewed-by: Coelho, Luciano <luciano.coelho@intel.com>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
x-iwlwifi-stack-dev: bbe30a5735f076cb7669cdb0e80a1eefce7afc45