iwlwifi: mvm: add initiator XVT capabilities
Add initiator flags for the range request.
The flags define the measurement process. for
example: which algorithm type to run and what are the
calibration values needs to be factored in.
Add debugfs control to the initiator flags. This shall
allow a user to override the default flag values and
control the measurement process.
Add range notification flags to enable to report
internal and partial algorithm products. for
example: t1t4_responder shall contain the result of the
algorithm on the responder side as was reported to the
initiator in the FTM action frame.

type=feature

Change-Id: Ibad4e5eef76cdc955012f5a6d22cae3f7c0d6689
Signed-off-by: Eyal Zolotov <eyal.zolotov@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/48520
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Coelho, Luciano <luciano.coelho@intel.com>
x-iwlwifi-stack-dev: 384825e45434e4e40fc76e13c3f5e64c416751c5