iwlwifi: mvm: add a TODO to clean up TOF debugfs entries for upstream
Some (or all?) TOF debugfs entries use a trigger string to actually
send the command to the firmware.  It should not be like this and must
be changed before upstreaming.  Add a big TODO to remind us of doing
that.

type=maintenance

Change-Id: Ifa00bf58d3b560fef82686faeee7fb97efb081b1
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/50244
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
x-iwlwifi-stack-dev: 1003794cecc1856891aee4dc210d36d425f248f1