[BUGFIX] iwlwifi: mvm: update TX queue before making a copy of the skb
Off-channel action frames (such as ANQP frames) must be sent either on
the AUX queue or on the offchannel queue, otherwise the firmware will
cause a SYSASSERT.

In the current implementation, the queue to be used is correctly set in
the original skb, but this is done after it is copied.  Thus the copy
remains with the original, incorrect queue.

Fix this by setting the queue in the original skb before copying it.

type=bugfix
bug=not-tracked
fixes=I6027f3631ac30b7beda10679e7563c9f6bdd3d26

Change-Id: I2e20d78614810678d7057e3e3a0931f254620756
Signed-off-by: Beni Lev <beni.lev@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/50616
Reviewed-by: Coelho, Luciano <luciano.coelho@intel.com>
Tested-by: Coelho, Luciano <luciano.coelho@intel.com>
x-iwlwifi-stack-dev: 6accedb32dae7b7c78dfcde6cae7778e94c52b07