iwlwifi: mvm: support MFUART dump in case of MFUART assert
In case of a MFUART assert, get a notification from the fw
that consists of the assert id and debug data.

The notification may be divided to multiple chunks, depending
on the size of the debug data sent to the driver, which would
be up to 1KB.

Get the notification, and if the debug info flag is enabled,
print the debug data to the dmesg.

type=feature

Change-Id: I6a38352dbe66e546bdab979e7fb1fc0cfe12c7d2
Signed-off-by: Golan Ben-Ami <golan.ben.ami@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/73040
Reviewed-by: Coelho, Luciano <luciano.coelho@intel.com>
Reviewed-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
x-iwlwifi-stack-dev: a39b743038c3be04fc94c781f626a2681544cb66