iwlwifi: move to TVQM mode
In TVQM firmware returns the value of the queue ID and code
should accept it.
The TX queue config API was changed. Move to new API.
This has to be done in parallel in mvm and pcie.
Do not move yet to 512 queues since there are some opens
with enabling it.

type=feature

Change-Id: I16db776e17d2ded041aafc4e0f60e5fb6cff2407
Signed-off-by: Sara Sharon <sara.sharon@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/87257
Reviewed-by: Coelho, Luciano <luciano.coelho@intel.com>
Tested-by: Coelho, Luciano <luciano.coelho@intel.com>
x-iwlwifi-stack-dev: 63fb7206976951b5e24dda27b395daf476620353