[BUGFIX] iwlwifi: mvm: fix rogue AP detection for MQ [NOUPSTREAM]
In MQ, the tpt we demand for declaring that the AP failed
AP detection is more than 22 percent of the phy rate. So if
the traffic is not more than 22 percent of the phy rate, the
AP can't fail AP detection, and then there is no reason for
disconnecting from it and cancelling U-APSD.

Flip the inequation mark so it would reflect this logic.

type=bugfix
bug=JIRA:WIFILNX-181
fixes=I963dac6c112579c286fd2fa81317e529452145c1

Change-Id: I9844b2bb651ce14b2cabae7b11386d1151b8881d
Signed-off-by: Golan Ben-Ami <golan.ben.ami@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/89341
Reviewed-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Sharon, Sara <sara.sharon@intel.com>
Reviewed-by: Berg, Johannes <johannes.berg@intel.com>
x-iwlwifi-stack-dev: 28907937da0357c1c0809b50caa559b2729fdb34