[BUGFIX] iwlwifi: mvm: move number of words calculation in MFUART to a variable [SQUASH]
The type of the value in the offset calculation was different for
32-bit and 64-bit architectures, so the format string would cause a
warning in 32-bit archs.  To avoid this, and to make the code look a
bit cleaner, move the number of words calculation to an integer
variable and use that in the loop.

type=bugfix
bug=not-tracked
fixes=I6a38352dbe66e546bdab979e7fb1fc0cfe12c7d2

Change-Id: I8c988f57f26ecedcffe69e19f17a68eda720cdd5
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/89999
Reviewed-by: Ben Ami, Golan <golan.ben.ami@intel.com>
Reviewed-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
x-iwlwifi-stack-dev: e6cd611484ed2556f458a44d34d580dab071ff13