[BUGFIX] iwlwifi: mvm: fix TX command creation [SQUASH]
My previous patch had a few issues:

It moved the memcpy of the header to be before the sequence control
assignment that modifies the header.
This may result with the sequence control of the command not
aligned with that of the command.
For pre-a000 devices, in case the queue index doesn't match the
sequence control, it will throw the scheduler completely out of
sync.

In addition, it broke support of transmitting unsecured frames to
a secured station for a000 devices.
mac80211 won't set the key for such frames, and for a000 devices
we need to check if key is set and pass it on to FW.

Fix all those issues.

Remove a redundant __packed added to an enum, and add a TODO
for AMSDU while at it.

type=bugfix
fixes=I78bb50e441d05d97533fedafa94c7745c48d45fb
bug=not-tracked

Change-Id: If967ec6ef9c6b9f46eec3a286800f36737de6cab
Reported-by: Johannes Berg <johannes.berg@intel.com>
Signed-off-by: Sara Sharon <sara.sharon@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/95340
Reviewed-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Grumbach, Emmanuel <emmanuel.grumbach@intel.com>
x-iwlwifi-stack-dev: 5ab2d691a2183e9db7ea2a4aa21b6711300e5f6e