iwlwifi: fmac: introduce 9000-series Kconfig option
In order to be able to prune out 9000-series support
from the driver later, for documentation purposes,
introduce a Kconfig option for this.

For upstream, the Kconfig option should probably be
removed in a way that only the ifdefs are removed
and the code is left, i.e. assuming it's set.

type=feature

Change-Id: I66e24aec61cc9af0265ac8161b36416026cdd74f
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/96842
Reviewed-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Grumbach, Emmanuel <emmanuel.grumbach@intel.com>
x-iwlwifi-stack-dev: ca5171b6c11b533f8bd6471dc638de03c3beea93