[BUGFIX] mac80211: don't parse encrypted management frames in ieee80211_frame_acked
ieee80211_frame_acked is called when a frame is acked by
the peer. In case this is a management frame, we check
if this an SMPS frame, in which case we can update our
antenna configuration.

When we parse the management frame we look at the category
in case it is an action frame. That byte sits after the IV
in case the frame was encrypted. This means that if the
frame was encrypted, we basically look at the IV instead
of looking at the category. It is then theorically
possible that we think that an SMPS action frame was acked
where really we had another frame that was encrypted.

Since the only management frame whose ack needs to be
tracked is the SMPS action frame, and that frame is not
a robust management frame, it will never be encrypted.
The easiest way to fix this problem is then to not look
at frames that were encrypted.

type=bugfix
fixes=unknown
bug=not-tracked

Change-Id: Iaa9cb3a33f8d4c420b3fe27f8e42c5c345d92634
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/104878
Reviewed-by: Coelho, Luciano <luciano.coelho@intel.com>
Tested-by: Coelho, Luciano <luciano.coelho@intel.com>
x-iwlwifi-stack-dev: 5d8a3b315c4ca7b9d31da2e70e48719e690c5240